Fix cyclic import and Pin class type
