+++ src/adafruit_blinka/microcontroller/stm32/stm32f405/pin.py
@@ -3,7 +3,7 @@
 # SPDX-License-Identifier: MIT
 """STM32F405 pins"""
 
-from microcontroller import Pin
+from ...generic_micropython import Pin
 
 A0 = Pin("A0")
 A1 = Pin("A1")
